producer_dir/requirements.txt
pandas==1.3.5
kafka-python==2.0.2


 